+++ common/src/main/java/me/ultrusmods/extrasponges/ExtraSpongesCommon.java
@@ -1,0 +1,8 @@
+package me.ultrusmods.extrasponges;
+
+public class ExtraSpongesCommon {
+
+    public static void init() {
+
+    }
+}