+++ common/src/main/java/me/ultrusmods/extrasponges/block/ExtraWetSpongeBlock.java
@@ -1,0 +1,68 @@
+package me.ultrusmods.extrasponges.block;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.core.particles.ParticleTypes;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+
+public class ExtraWetSpongeBlock extends Block {
+    private final BlockState originalBlock;
+    public ExtraWetSpongeBlock(Properties settings, BlockState originalBlock) {
+        super(settings);
+        this.originalBlock = originalBlock;
+    }
+
+    @Override
+    public void onPlace(BlockState state, Level level, BlockPos pos, BlockState oldState, boolean notify) {
+        if (level.dimensionType().ultraWarm()) {
+            level.setBlock(pos, originalBlock, 3);
+            level.globalLevelEvent(2009, pos, 0);
+            level.playSound(null, pos, SoundEvents.FIRE_EXTINGUISH, SoundSource.BLOCKS, 1.0F, (1.0F + level.getRandom().nextFloat() * 0.2F) * 0.7F);
+        }
+
+    }
+
+    @Override
+    public void animateTick(BlockState state, Level level, BlockPos pos, RandomSource random) {
+        Direction direction = Direction.getRandom(random);
+        if (direction != Direction.UP) {
+            BlockPos blockPos = pos.relative(direction);
+            BlockState blockState = level.getBlockState(blockPos);
+            if (!state.canOcclude() || !blockState.isFaceSturdy(level, blockPos, direction.getOpposite())) {
+                double d = pos.getX();
+                double e = pos.getY();
+                double f = pos.getZ();
+                if (direction == Direction.DOWN) {
+                    e -= 0.05D;
+                    d += random.nextDouble();
+                    f += random.nextDouble();
+                } else {
+                    e += random.nextDouble() * 0.8D;
+                    if (direction.getAxis() == Direction.Axis.X) {
+                        f += random.nextDouble();
+                        if (direction == Direction.EAST) {
+                            ++d;
+                        } else {
+                            d += 0.05D;
+                        }
+                    } else {
+                        d += random.nextDouble();
+                        if (direction == Direction.SOUTH) {
+                            ++f;
+                        } else {
+                            f += 0.05D;
+                        }
+                    }
+                }
+
+                level.addParticle(ParticleTypes.DRIPPING_WATER, d, e, f, 0.0D, 0.0D, 0.0D);
+            }
+        }
+    }
+
+}