+++ fabric/src/main/java/me/ultrusmods/extrasponges/data/ExtraSpongesDataGenerator.java
@@ -1,0 +1,16 @@
+package me.ultrusmods.extrasponges.data;
+
+import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
+
+public class ExtraSpongesDataGenerator implements DataGeneratorEntrypoint {
+    @Override
+    public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
+        var pack = fabricDataGenerator.createPack();
+
+        pack.addProvider(ExtraSpongesModelGenerator::new);
+        pack.addProvider(ExtraSpongesLootGenerator::new);
+        pack.addProvider(ExtraSpongesRecipeGenerator::new);
+
+    }
+}