+++ fabric/src/main/java/me/ultrusmods/extrasponges/data/ExtraSpongesLootGenerator.java
@@ -1,0 +1,40 @@
+package me.ultrusmods.extrasponges.data;
+
+import me.ultrusmods.extrasponges.register.BlockRegistry;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricBlockLootTableProvider;
+import net.fabricmc.fabric.api.datagen.v1.provider.SimpleFabricLootTableProvider;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.level.storage.loot.LootTable;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParamSet;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
+
+import java.util.concurrent.CompletableFuture;
+import java.util.function.BiConsumer;
+
+public class ExtraSpongesLootGenerator extends FabricBlockLootTableProvider {
+
+
+    protected ExtraSpongesLootGenerator(FabricDataOutput dataOutput, CompletableFuture<HolderLookup.Provider> registryLookup) {
+        super(dataOutput, registryLookup);
+    }
+
+    @Override
+    public void generate() {
+        this.dropSelf(BlockRegistry.IRON_SPONGE);
+        this.dropSelf(BlockRegistry.GOLD_SPONGE);
+        this.dropSelf(BlockRegistry.DIAMOND_SPONGE);
+        this.dropSelf(BlockRegistry.NETHERITE_SPONGE);
+        this.dropSelf(BlockRegistry.WET_IRON_SPONGE);
+        this.dropSelf(BlockRegistry.WET_GOLD_SPONGE);
+        this.dropSelf(BlockRegistry.WET_DIAMOND_SPONGE);
+        this.dropSelf(BlockRegistry.WET_NETHERITE_SPONGE);
+        this.dropSelf(BlockRegistry.LAVA_SPONGE);
+        this.dropSelf(BlockRegistry.WET_LAVA_SPONGE);
+
+
+    }
+
+
+}