+++ fabric/src/main/java/me/ultrusmods/extrasponges/platform/FabricPlatformHelper.java
@@ -1,0 +1,32 @@
+package me.ultrusmods.extrasponges.platform;
+
+import me.ultrusmods.extrasponges.platform.services.IPlatformHelper;
+import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
+import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.world.item.CreativeModeTab;
+
+public class FabricPlatformHelper implements IPlatformHelper {
+
+    @Override
+    public String getPlatformName() {
+        return "Fabric";
+    }
+
+    @Override
+    public boolean isModLoaded(String modId) {
+
+        return FabricLoader.getInstance().isModLoaded(modId);
+    }
+
+    @Override
+    public boolean isDevelopmentEnvironment() {
+
+        return FabricLoader.getInstance().isDevelopmentEnvironment();
+    }
+
+    @Override
+    public CreativeModeTab.Builder getCreativeTab() {
+        return FabricItemGroup.builder();
+
+    }
+}